# stewardess_openapi3.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Stewardess MCP Service",
    "description": "Local source-code repository tools for AI agents. Provides file reading, search, editing, git, and command execution.",
    "version": "v1"
  },
  "servers": [
    {
      "url": "http://office:55702"
    }
  ],
  "paths": {
...
    "/api/repository/directory": {
      "get": {
        "operationId": "Repository_ListDirectory",
        "summary": "Lists the immediate contents of a directory.",
        "description": "Lists the immediate contents of a directory.",
        "tags": [
          "Repository"
        ],
        "parameters": [
          {
            "name": "path",
            "in": "query",
            "required": false,
            "description": "",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "includeBlocked",
            "in": "query",
            "required": false,
            "description": "",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "namePattern",
            "in": "query",
            "required": false,
            "description": "",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/api/repository/tree": {
      "get": {
        "operationId": "Repository_ListTree",
        "summary": "Returns a recursive directory tree.",
        "description": "Returns a recursive directory tree.",
        "tags": [
          "Repository"
        ],
        "parameters": [
          {
            "name": "path",
            "in": "query",
            "required": false,
            "description": "",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "maxDepth",
            "in": "query",
            "required": false,
            "description": "",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "includeBlocked",
            "in": "query",
            "required": false,
            "description": "",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "directoriesOnly",
            "in": "query",
            "required": false,
            "description": "",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/api/repository/exists": {
      "get": {
        "operationId": "Repository_PathExists",
        "summary": "Checks whether a path exists inside the repository.",
        "description": "Checks whether a path exists inside the repository.",
        "tags": [
          "Repository"
        ],
        "parameters": [
          {
            "name": "path",
            "in": "query",
            "required": true,
            "description": "",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/api/repository/metadata": {
      "get": {
        "operationId": "Repository_GetMetadata",
        "summary": "Returns detailed metadata for a file or directory.",
        "description": "Returns detailed metadata for a file or directory.",
        "tags": [
          "Repository"
        ],
        "parameters": [
          {
            "name": "path",
            "in": "query",
            "required": true,
            "description": "",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/search/text": {
      "post": {
        "operationId": "Search_SearchTextPost",
        "summary": "Searches for a literal string (POST variant for full options).",
        "description": "Searches for a literal string (POST variant for full options).",
        "tags": [
          "Search"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchTextRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/api/search/regex": {
      "post": {
        "operationId": "Search_SearchRegex",
        "summary": "Searches using a .NET regular expression pattern.",
        "description": "Searches using a .NET regular expression pattern.",
        "tags": [
          "Search"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchRegexRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/api/search/files": {
      "get": {
        "operationId": "Search_SearchFiles",
        "summary": "Finds files whose names match a wildcard or substring.",
        "description": "Finds files whose names match a wildcard or substring.",
        "tags": [
          "Search"
        ],
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "description": "",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "path",
            "in": "query",
            "required": false,
            "description": "",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "ignoreCase",
            "in": "query",
            "required": false,
            "description": "",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "fullPath",
            "in": "query",
            "required": false,
            "description": "",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "maxResults",
            "in": "query",
            "required": false,
            "description": "",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/api/search/symbol": {
      "post": {
        "operationId": "Search_SearchSymbol",
        "summary": "Best-effort symbol search via text heuristics.",
        "description": "Best-effort symbol search via text heuristics.",
        "tags": [
          "Search"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchSymbolRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/api/search/references": {
      "post": {
        "operationId": "Search_FindReferences",
        "summary": "Finds textual references to an identifier across the repository.",
        "description": "Finds textual references to an identifier across the repository.",
        "tags": [
          "Search"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FindReferencesRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "SearchTextRequest": {
        "description": "Request to search for literal text across repository files.",
        "type": "object",
        "properties": {
          "query": {
            "description": "Literal text to search for.",
            "type": "string"
          },
          "ignoreCase": {
            "description": "Case-insensitive search (default true).",
            "type": "boolean"
          },
          "wholeWord": {
            "description": "Match whole words only.",
            "type": "boolean"
          },
          "searchPath": {
            "description": "Restrict search to this subdirectory (relative path; empty = whole repo).",
            "type": "string"
          },
          "extensions": {
            "description": "Restrict to these file extensions (e.g. \".cs\", \".json\").  Empty = all.",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "maxResults": {
            "format": "int32",
            "description": "Maximum number of results to return.  Capped at server limit.",
            "type": "integer"
          },
          "sortByPath": {
            "description": "When true, results are sorted by file path.",
            "type": "boolean"
          },
          "contextLinesBefore": {
            "format": "int32",
            "description": "Number of context lines to include before each match.",
            "type": "integer"
          },
          "contextLinesAfter": {
            "format": "int32",
            "description": "Number of context lines to include after each match.",
            "type": "integer"
          }
        }
      },
      "SearchRegexRequest": {
        "description": "Request to search repository files using a .NET regular expression.",
        "type": "object",
        "properties": {
          "pattern": {
            "description": ".NET regular expression pattern.",
            "type": "string"
          },
          "ignoreCase": {
            "description": "Case-insensitive (default true).",
            "type": "boolean"
          },
          "multiline": {
            "description": "Apply the pattern in multiline mode (^ and $ match line boundaries).",
            "type": "boolean"
          },
          "searchPath": {
            "description": "Restrict search to this subdirectory (relative path; empty = whole repo).",
            "type": "string"
          },
          "extensions": {
            "description": "Restrict to these file extensions (e.g. \".cs\", \".json\").  Empty = all.",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "maxResults": {
            "format": "int32",
            "description": "Maximum number of results to return.  Capped at server limit.",
            "type": "integer"
          },
          "sortByPath": {
            "description": "When true, results are sorted by file path.",
            "type": "boolean"
          },
          "contextLinesBefore": {
            "format": "int32",
            "description": "Number of context lines to include before each match.",
            "type": "integer"
          },
          "contextLinesAfter": {
            "format": "int32",
            "description": "Number of context lines to include after each match.",
            "type": "integer"
          }
        }
      },
...
      "SearchSymbolRequest": {
        "description": "Best-effort symbol search using text heuristics.\r\nFinds class names, method signatures, interface names, etc.",
        "type": "object",
        "properties": {
          "symbolName": {
            "description": "Symbol name or partial name to search for.",
            "type": "string"
          },
          "symbolKind": {
            "description": "Symbol kind filter: \"class\", \"interface\", \"method\", \"property\", \"field\", \"\" (all).",
            "type": "string"
          },
          "ignoreCase": {
            "description": "When true, performs a case-insensitive name comparison.",
            "type": "boolean"
          },
          "searchPath": {
            "description": "Restrict search to this subdirectory (relative path; empty = whole repo).",
            "type": "string"
          },
          "extensions": {
            "description": "Restrict to these file extensions (e.g. \".cs\", \".json\").  Empty = all.",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "maxResults": {
            "format": "int32",
            "description": "Maximum number of results to return.  Capped at server limit.",
            "type": "integer"
          },
          "sortByPath": {
            "description": "When true, results are sorted by file path.",
            "type": "boolean"
          },
          "contextLinesBefore": {
            "format": "int32",
            "description": "Number of context lines to include before each match.",
            "type": "integer"
          },
          "contextLinesAfter": {
            "format": "int32",
            "description": "Number of context lines to include after each match.",
            "type": "integer"
          }
        }
      },
      "FindReferencesRequest": {
        "description": "Request to find textual usages of an identifier across the repository.",
        "type": "object",
        "properties": {
          "identifierName": {
            "description": "Identifier name whose usages (textual references) should be found.",
            "type": "string"
          },
          "ignoreCase": {
            "description": "When true, performs a case-insensitive comparison.",
            "type": "boolean"
          },
          "searchPath": {
            "description": "Restrict search to this subdirectory (relative path; empty = whole repo).",
            "type": "string"
          },
          "extensions": {
            "description": "Restrict to these file extensions (e.g. \".cs\", \".json\").  Empty = all.",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "maxResults": {
            "format": "int32",
            "description": "Maximum number of results to return.  Capped at server limit.",
            "type": "integer"
          },
          "sortByPath": {
            "description": "When true, results are sorted by file path.",
            "type": "boolean"
          },
          "contextLinesBefore": {
            "format": "int32",
            "description": "Number of context lines to include before each match.",
            "type": "integer"
          },
          "contextLinesAfter": {
            "format": "int32",
            "description": "Number of context lines to include after each match.",
            "type": "integer"
          }
        }
      }
    }
  }
}